Move top level package to internal (#5)
