+++ internal/server/templates/index.html
@@ -90,7 +90,6 @@ function updateFormAction() {
 	name="data-iframe"
 	class="data-iframe"
 	>
-
 </div>
 </body>
 </html>
